Refactored game logic into separate class. Added asset field to piece classes.
